Commit ÁLvaro tras juntar todo loq ue trabajamos ayer: RulesConfig, FeePOayemtnService, ficheros drl (ahora solo hay uno), han sido cambiados.

+++ Acme-Six-Pack/src/main/java/es/us/lsi/dp/controllers/core/BaseController.java
@@ -197,7 +197,7 @@ public abstract class BaseController {
 		DecimalFormat numberFormat = new DecimalFormat("##.##", decimalFormatSymbols);
 		numberFormat.setPositivePrefix(prefix);
 		numberFormat.setPositiveSuffix(suffix);
-		binder.registerCustomEditor(BigDecimal.class, "fee", new CustomNumberEditor(BigDecimal.class, numberFormat, true));
+		binder.registerCustomEditor(Double.class, "fee", new CustomNumberEditor(Double.class, numberFormat, true));
 		
 		//CustomCurrencyFormat  customCurrencyFormat = new CustomCurrencyFormat("", BigDecimal.class, "");
 		//binder.registerCustomEditor(customCurrencyFormat.getType(), new CustomNumberEditor(customCurrencyFormat.getType(),customCurrencyFormat.getFormat(), true));
